+++ bootstrap-scripts/package-install.sh
@@ -31,7 +31,10 @@ NTP_SERVERS=$(echo "$NTP_SERVERS" | sed -e 's|[]"'\''\[ ]||g')
 iptables -A LOGGING -d  $NTP_SERVERS -j ACCEPT # NTP server
 fi
 if [ "x$networkCidr" != "x" ]; then
-iptables -A LOGGING -d  ${networkCidr} -j ACCEPT # PNDA network for AWS
+iptables -A LOGGING -d  ${networkCidr} -j ACCEPT
+fi
+if [ "x$privateSubnetCidr" != "x" ]; then
+iptables -A LOGGING -d  ${privateSubnetCidr} -j ACCEPT
 fi
 iptables -A LOGGING -j REJECT --reject-with icmp-net-unreachable
 iptables-save > /etc/iptables.conf
@@ -87,13 +90,3 @@ cat << EOF > /root/.pydistutils.cfg
 [easy_install]
 index_url=$PIP_INDEX_URL
 EOF
-
-if [ "x$ADD_ONLINE_REPOS" == "xYES" ]; then
-cat << EOF >> /etc/pip.conf
-extra-index-url=https://pypi.python.org/simple/
-EOF
-cat << EOF >> /root/.pydistutils.cfg
-find_links=https://pypi.python.org/simple/
-EOF
-fi
-